refactor: split ITestJsonFormatting into per-interface files, add respective implementation files
