+++ tests/FunctionalTests/CMakeLists.txt
@@ -25,10 +25,14 @@ option(TEST_OPTIONALS    "Enable optional parameters tests"     ON)
 option(TEST_PRIMITIVES   "Enable primitive types tests"         ON)
 option(TEST_RESTRICTIONS "Enable restrict annotation tests"     ON) 
 option(TEST_STRUCTS      "Enable POD struct tests"              ON)
-option(TEST_ENCODING_MAC "Enable @encode:mac annotation tests"           ON)
-option(TEST_LENGTH_MODES "Enable @length:void/@length:return tests"      ON)
-option(TEST_JSON_SHAPE   "Enable @wrapped/@extract shape tests"          ON)
-option(TEST_JSON_FORMATTING "Enable @text/@compliant/@uncompliant tests"    ON)
+option(TEST_ENCODING_MAC         "Enable @encode:mac annotation tests"           ON)
+option(TEST_LENGTH_MODES         "Enable @length:void/@length:return tests"      ON)
+option(TEST_JSON_SHAPE           "Enable @wrapped/@extract shape tests"          ON)
+option(TEST_JSON_TEXT_KEEP       "Enable @text:keep naming tests"                ON)
+option(TEST_JSON_TEXT_CASE       "Enable @text:legacy case convention tests"     ON)
+option(TEST_JSON_COMPLIANT       "Enable @compliant format tests"                ON)
+option(TEST_JSON_UNCOMPLIANT_EXT "Enable @uncompliant:extended format tests"     ON)
+option(TEST_JSON_UNCOMPLIANT_COL "Enable @uncompliant:collapsed format tests"    ON)
 
 set(CMAKE_MODULE_PATH "${CMAKE_BINARY_DIR}" CACHE BOOL "" FORCE)
