+++ tests/FunctionalTests/comrpc/CMakeLists.txt
@@ -71,7 +71,7 @@ if(TEST_JSON_SHAPE)
     target_sources(ComRpcFunctionalTests PRIVATE tests/TestJsonShape.cpp)
 endif()
 
-if(TEST_JSON_FORMATTING)
+if(TEST_JSON_TEXT_KEEP OR TEST_JSON_TEXT_CASE OR TEST_JSON_COMPLIANT OR TEST_JSON_UNCOMPLIANT_EXT OR TEST_JSON_UNCOMPLIANT_COL)
     target_sources(ComRpcFunctionalTests PRIVATE tests/TestJsonFormatting.cpp)
 endif()
